sbc: Add encoder_state to analysis functions
Until now, SIMD analysis used to process 4 blocks of 8 samples at a
time. This was implemented using two constant tables: odd and even. This
mean we can only process 4, 8, 12, or 16 blocks par SBC packets.
mSBC requires 15 blocks, so to be able to analyse 1 block, it will be
necessary to know if we are processing an odd or even block. This will
be done with a new member to encoder_state.